# perf/fix: use a parser for gradients.