- Don't use the regex for gradients. The regex had the potential to slow down DR a lot and even had the ability to freeze the browser all together.
- Having less Regex is good and add more readability.
- Performance improvements:
 System Info:
Chromium 94.0.4606.81 & Firefox Nightly 95.0a1 (2021-10-12) (64-bit) 
(Arch) Linux 5.14.11 (zen-kernel).

Extreme-case:
- `"linear-gradient(217deg, rgba(255,0,0,.8), rgba(255,0,0,0) 70.71%), linear-gradient(127deg, rgba(0,255,0,.8), rgba(0,255,0,0) 70.71%), linear-gradient(336deg, rgba(0,0,255,.8), rgba(0,0,255,0) 70.71%);"`

Before:
- Chromium: 693,666 ops/sec ±0.23% (97 runs sampled)
- Firefox: 647,798 ops/sec ±0.36% (95 runs sampled)

After:
- Chromium: 1,695,039 ops/sec ±0.89% (95 runs sampled)
- Firefox 695,486 ops/sec ±0.51% (97 runs sampled)


Possible input(no gradients):
- `url("../../media/examples/star.png"), url("../../media/examples/lizard.png");`

Before:
- Chromium: 2,413,792 ops/sec ±0.26% (100 runs sampled)
- Firefox: 2,105,840 ops/sec ±0.61% (98 runs sampled)

After:
- Chromium: 56,380,341 ops/sec ±0.52% (96 runs sampled)
- Firefox: 10,010,921 ops/sec ±0.91% (95 runs sampled)


Possible input(gradient + background-image):
- `linear-gradient(rgba(0, 0, 255, 0.5), rgba(255, 255, 0, 0.5)), url("../../media/examples/lizard.png");`

Before:
- Chromium: 1,354,331 ops/sec ±0.53% (101 runs sampled)
- Firefox: 1,187,130 ops/sec ±0.75% (100 runs sampled)

After:
- Chromium: 4,289,281 ops/sec ±1.05% (98 runs sampled)
- Firefox: 1,819,759 ops/sec ±0.49% (98 runs sampled)


Possible input(simple radial):
- `radial-gradient(closest-side, #3f87a6, #ebf8e1, #f69d3c);`

Before:
- Chromium: 1,978,929 ops/sec ±1.13% (100 runs sampled)
- Firefox: 1,721,784 ops/sec ±0.72% (98 runs sampled)

After:
- Chromium: 6,019,032 ops/sec ±1.27% (93 runs sampled)
- Firefox: 2,512,294 ops/sec ±0.47% (96 runs sampled)